Adds module code organization
